// ebanking-frontend/src/app/model/account.model.ts
export interface AccountDetailsModel {
  accountId: number;
  balance: number;
  currentPage: number;
  totalPages: number;
  pageSize: number;
  accountOperationDTOS : AccountOperation[]
}

export interface AccountOperation{
  id: number;
  operationDate:Date;
  amount: number;
  type: string;
  description: string;
}